wip obtain delete resource
